requirements.txt
fastapi[all]
pyzmq